move `get_block_by_block_hash` in `madara_backend_client` move `transaction` in utils: conversion from starknet-api to blockifier
