+++ crates/client/rpc/src/lib.rs
@@ -5,7 +5,7 @@
 mod constants;
 mod errors;
 mod events;
-mod madara_backend_client;
+pub mod madara_backend_client;
 mod methods;
 mod types;
 pub mod utils;
@@ -18,8 +18,6 @@ use jsonrpsee::core::RpcResult;
 use jsonrpsee::proc_macros::rpc;
 use mc_db::DeoxysBackend;
 use mc_storage::OverrideHandle;
-use mp_block::DeoxysBlock;
-use mp_digest_log::find_starknet_block;
 use mp_felt::Felt252Wrapper;
 use mp_hashers::HasherT;
 use mp_types::block::{DBlockT, DHashT, DHeaderT};
@@ -28,7 +26,7 @@ use sc_network_sync::SyncingService;
 use sc_transaction_pool::{ChainApi, Pool};
 use serde::{Deserialize, Serialize};
 use serde_with::serde_as;
-use sp_api::{BlockT, ProvideRuntimeApi};
+use sp_api::ProvideRuntimeApi;
 use sp_arithmetic::traits::UniqueSaturatedInto;
 use sp_blockchain::HeaderBackend;
 use sp_core::H256;
@@ -45,6 +43,7 @@ use starknet_core::types::{
     SyncStatusType, Transaction, TransactionReceiptWithBlockInfo, TransactionStatus, TransactionTraceWithHash,
 };
 
+use crate::madara_backend_client::get_block_by_block_hash;
 use crate::methods::get_block::{
     get_block_with_tx_hashes_finalized, get_block_with_tx_hashes_pending, get_block_with_txs_finalized,
     get_block_with_txs_pending,
@@ -351,17 +350,4 @@ where
 
         Ok(rpc_state_diff)
     }
-}
-
-/// Returns the current Starknet block from the block header's digest
-pub fn get_block_by_block_hash<B, C>(client: &C, block_hash: <B as BlockT>::Hash) -> anyhow::Result<DeoxysBlock>
-where
-    B: BlockT,
-    C: HeaderBackend<B>,
-{
-    let header =
-        client.header(block_hash).ok().flatten().ok_or_else(|| anyhow::Error::msg("Failed to retrieve header"))?;
-    let digest = header.digest();
-    let block = find_starknet_block(digest)?;
-    Ok(block)
 }
