+++ crates/client/rpc/src/methods/read/get_transaction_by_hash.rs
@@ -16,7 +16,8 @@ use sp_blockchain::HeaderBackend;
 use starknet_core::types::{FieldElement, Transaction};
 
 use crate::errors::StarknetRpcApiError;
-use crate::{get_block_by_block_hash, Starknet};
+use crate::madara_backend_client::get_block_by_block_hash;
+use crate::Starknet;
 
 /// Get the details and status of a submitted transaction.
 ///
