+++ crates/client/rpc/src/utils/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod block;
 pub(crate) mod call_info;
 pub(crate) mod utils;
+pub(crate) mod transaction;
